edit controllers

+++ src/main/java/com/example/semestralka/controllers/UserController.java
@@ -2,16 +2,19 @@ package com.example.semestralka.controllers;
 
 import com.example.semestralka.controllers.util.RestUtils;
 import com.example.semestralka.exceptions.NotFoundException;
+import com.example.semestralka.exceptions.PersistenceException;
 import com.example.semestralka.model.Event;
 import com.example.semestralka.model.User;
 import com.example.semestralka.security.model.UserDetails;
 import com.example.semestralka.services.UserService;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -27,6 +30,7 @@ public class UserController {
         this.userService = userService;
     }
 
+    @PreAuthorize("(!#user.isAdmin() && anonymous) || hasRole('ROLE_ADMIN')")
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> register(@RequestBody User user) {
         userService.save(user);
@@ -34,42 +38,39 @@ public class UserController {
         return new ResponseEntity<>(headers, HttpStatus.CREATED);
     }
 
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_USER', 'ROLE_GUEST')")
     @GetMapping(value = "/current", produces = MediaType.APPLICATION_JSON_VALUE)
     public User getCurrent(Authentication auth) {
         assert auth.getPrincipal() instanceof UserDetails;
         return ((UserDetails) auth.getPrincipal()).getUser();
     }
 
-    @GetMapping("/favorites")
-    public ResponseEntity<List<Event>> getFavoritesByUserId(@RequestBody User user) {
-        List<Event> favorites = userService.getAllFavoriteEvents(user);
-        return new ResponseEntity<>(favorites, HttpStatus.OK);
-    }
-
-    @GetMapping("/favorites/upcoming")
-    public ResponseEntity<List<Event>> getUpcomingFavoritesByUserId(@RequestBody User user) {
-        List<Event> favorites = userService.getAllFavoriteUpcomingEvents(user);
-        return new ResponseEntity<>(favorites, HttpStatus.OK);
-    }
-
     @PostMapping("/edit")
-    public ResponseEntity<Void> updateUser(@RequestBody User user) {
-        if (userService.exists(user.getId())) {
-            userService.save(user);
-            final HttpHeaders headers = RestUtils.createLocationHeaderFromCurrentUri("/current");
-            return new ResponseEntity<>(headers, HttpStatus.OK);
-        } else {
-            throw NotFoundException.create("User", user.getId());
-        }
+    @PreAuthorize("hasRole('ROLE_USER')")
+    public ResponseEntity<Void> updateUser(@RequestBody User user, @AuthenticationPrincipal UserDetails userDetails) {
+        if (user.getId().equals(userDetails.getUser().getId())) {
+            if (userService.exists(user.getId())) {
+                userService.save(user);
+                final HttpHeaders headers = RestUtils.createLocationHeaderFromCurrentUri("/current");
+                return new ResponseEntity<>(headers, HttpStatus.OK);
+            } else {
+                throw NotFoundException.create("User", user.getId());
+            }
+        } else return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_USER')")
     @DeleteMapping("/{userId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable Integer userId) {
-        final User user = userService.find(userId);
-        if (user == null) {
-            throw NotFoundException.create("User", userId);
-        }
-        userService.delete(user);
+    public void deleteUser(@PathVariable Integer userId, @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"))
+                || userId.equals(userDetails.getUser().getId())) {
+            final User user = userService.find(userId);
+            if (user == null) {
+                throw NotFoundException.create("User", userId);
+            }
+            userService.delete(user);
+        } else throw NotFoundException.create("User", userId);
     }
+
 }
